+++ blueprint/src/main/kotlin/no/eolseng/pg6102/blueprint/dto/BlueprintDto.kt
@@ -14,7 +14,7 @@ data class BlueprintDto(
         var description: String? = null,
 
         @get:ApiModelProperty("The value of the blueprint")
-        var value: Long? = null
+        var value: Int? = null
 
 )
